Rename `time` utils file
`addLeadingZero` function is not really realted to time - it's more realted to
numbers. So we are changing the name of the util file:

time.ts -> number.ts
