+++ src/utils/index.ts
@@ -1,2 +1,2 @@
 export * from "./date"
-export * from "./time"
+export * from "./number"
